requirements.txt
blinker==1.4
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
Django==1.11.4
django-filter==1.0.4
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.3.1
django-test-plus==1.0.18
djangorestframework==3.6.3
idna==2.5
mongoengine==0.13.0
py==1.4.34
pymongo==3.5.0
pytest==3.2.1
pytz==2017.2
requests==2.18.3
six==1.10.0
urllib3==1.22
